+++ PassPalette/ZXCVBN/BBSequenceMatcher.m
@@ -38,14 +38,14 @@ static NSDictionary *g_sequences;
         NSDictionary *sequences = [BBSequenceMatcher sequences];
         for (NSString *candidateName in sequences) {
             NSString *candidateSequence = [sequences objectForKey:candidateName];
-            int iPosition = [candidateSequence rangeOfString:[password substringWithRange:NSMakeRange(i, 1)]].location;
-            int jPosition = NSNotFound;
+            NSUInteger iPosition = [candidateSequence rangeOfString:[password substringWithRange:NSMakeRange(i, 1)]].location;
+            NSUInteger jPosition = NSNotFound;
             if (j < password.length) {
                 jPosition = [candidateSequence rangeOfString:[password substringWithRange:NSMakeRange(j, 1)]].location;
             }
             
             if (iPosition != NSNotFound && jPosition != NSNotFound) {
-                int direction = jPosition - iPosition;
+                int direction = (int)jPosition - (int)iPosition;
                 if (abs(direction) == 1) {
                     sequence = candidateSequence;
                     sequenceName = candidateName;
@@ -57,14 +57,14 @@ static NSDictionary *g_sequences;
         
         if (sequence) {
             while (1) {
-                int previousPosition = NSNotFound, currentPosition = NSNotFound;
+                NSUInteger previousPosition = NSNotFound, currentPosition = NSNotFound;
                 if (j < password.length) {
                     NSString *previousChar = [password substringWithRange:NSMakeRange(j - 1, 1)];
                     NSString *currentChar = [password substringWithRange:NSMakeRange(j, 1)];
                     previousPosition = [sequence rangeOfString:previousChar].location;
                     currentPosition = [sequence rangeOfString:currentChar].location;
                 }
-                if (j == password.length || currentPosition - previousPosition != sequenceDirection) {
+                if (j == password.length || (int)currentPosition - (int)previousPosition != sequenceDirection) {
                     if (j - i > 2) {
                         BBPattern *match = [[BBPattern alloc] init];
                         match.type = BBPatternTypeSequence;
